Separate out for sync tests

+++ hosted/eks/p1/p1_provisioning_test.go
@@ -17,7 +17,6 @@ import (
 
 var _ = Describe("P1Provisioning", func() {
 	var cluster *management.Cluster
-	loggingTypes := []string{"api", "audit", "authenticator", "controllerManager", "scheduler"}
 
 	var _ = BeforeEach(func() {
 		var err error
@@ -29,9 +28,11 @@ var _ = Describe("P1Provisioning", func() {
 	Context("Provisioning/Editing a cluster with invalid config", func() {
 
 		AfterEach(func() {
-			if ctx.ClusterCleanup {
-				err := helper.DeleteEKSHostCluster(cluster, ctx.RancherAdminClient)
-				Expect(err).To(BeNil())
+			if ctx.ClusterCleanup && cluster != nil {
+				if cluster != nil {
+					err := helper.DeleteEKSHostCluster(cluster, ctx.RancherAdminClient)
+					Expect(err).To(BeNil())
+				}
 			}
 		})
 
@@ -141,7 +142,7 @@ var _ = Describe("P1Provisioning", func() {
 		})
 
 		AfterEach(func() {
-			if ctx.ClusterCleanup {
+			if ctx.ClusterCleanup && cluster != nil {
 				err := helper.DeleteEKSHostCluster(cluster, ctx.RancherAdminClient)
 				Expect(err).To(BeNil())
 			} else {
@@ -206,38 +207,6 @@ var _ = Describe("P1Provisioning", func() {
 			testCaseID = 148
 			updateClusterInUpdatingState(cluster, ctx.RancherAdminClient)
 		})
-
-		It("Upgrade k8s version of cluster from EKS and verify it is synced back to Rancher", func() {
-			testCaseID = 159
-
-			By("upgrading the ControlPlane & NodeGroup", func() {
-				syncK8sVersionUpgradeCheck(cluster, ctx.RancherAdminClient, true)
-			})
-		})
-
-		It("Sync from Rancher to AWS console after a sync from AWS console to Rancher", func() {
-			testCaseID = 157
-
-			var err error
-			currentNodeGroupNumber := len(cluster.EKSConfig.NodeGroups)
-			By("upgrading control plane", func() {
-				syncK8sVersionUpgradeCheck(cluster, ctx.RancherAdminClient, false)
-			})
-
-			By("adding a NodeGroup", func() {
-				cluster, err = helper.AddNodeGroup(cluster, 1, ctx.RancherAdminClient, true, true)
-				Expect(err).To(BeNil())
-			})
-
-			By("Adding the LoggingTypes", func() {
-				cluster, err = helper.UpdateLogging(cluster, ctx.RancherAdminClient, loggingTypes, true)
-				Expect(err).To(BeNil())
-			})
-
-			// Check if the desired config is set correctly
-			Expect(len(cluster.EKSConfig.NodeGroups)).Should(BeNumerically("==", currentNodeGroupNumber+1))
-			Expect(*cluster.EKSConfig.LoggingTypes).Should(HaveExactElements(loggingTypes))
-		})
 	})
 
 	When("a cluster is created", func() {
@@ -251,7 +220,7 @@ var _ = Describe("P1Provisioning", func() {
 		})
 
 		AfterEach(func() {
-			if ctx.ClusterCleanup {
+			if ctx.ClusterCleanup && cluster != nil {
 				err := helper.DeleteEKSHostCluster(cluster, ctx.RancherAdminClient)
 				Expect(err).To(BeNil())
 			} else {
@@ -263,6 +232,7 @@ var _ = Describe("P1Provisioning", func() {
 			testCaseID = 128
 
 			var err error
+			loggingTypes := []string{"api", "audit", "authenticator", "controllerManager", "scheduler"}
 			By("Adding the LoggingTypes", func() {
 				cluster, err = helper.UpdateLogging(cluster, ctx.RancherAdminClient, loggingTypes, true)
 				Expect(err).To(BeNil())
